# [bugfix] Replace custom direction filter for loopback interfaces with BPF filter