--- a/link/link.go
+++ b/link/link.go
@@ -92,8 +92,9 @@ func (l Type) IPHeaderOffset() byte {
 // BPFFilter returns the link / interface specific raw BPF instructions to filter for valid packets only
 func (l Type) BPFFilter() func(snapLen int) []bpf.RawInstruction {
 	switch l {
-	case TypeEthernet,
-		TypeLoopback:
+	case TypeLoopback:
+		return bpfInstructionsLinkTypeLoopback
+	case TypeEthernet:
 		return bpfInstructionsLinkTypeEther
 	case TypePPP,
 		TypeIP6IP6,
@@ -110,16 +111,6 @@ func (l Type) BPFFilter() func(snapLen int) []bpf.RawInstruction {
 // Link denotes a link, i.e. an interface (wrapped) and its link type
 type Link struct {
 	Interface
-
-	filterMask byte
-}
-
-// WithPacketFilterMask sets / enables a packet type filter (masking) that will be applied
-// during capture
-func WithPacketFilterMask(mask byte) func(l *Link) {
-	return func(l *Link) {
-		l.filterMask |= mask
-	}
 }
 
 // New instantiates a new link / interface
@@ -151,8 +142,7 @@ func New(ifName string, opts ...func(*Link)) (link *Link, err error) {
 	}
 
 	link = &Link{
-		Interface:  iface,
-		filterMask: calculateInitialFilterMask(iface),
+		Interface: iface,
 	}
 
 	// Apply functional options, if any
@@ -166,11 +156,6 @@ func New(ifName string, opts ...func(*Link)) (link *Link, err error) {
 // IsUp returns if a link / interface is up
 func (l *Link) IsUp() (bool, error) {
 	return l.Interface.IsUp()
-}
-
-// FilterMask returns the packet type filter for this link / interface
-func (l *Link) FilterMask() byte {
-	return l.filterMask
 }
 
 // FindAllLinks retrieves all system network interfaces and their link type
@@ -191,16 +176,4 @@ func FindAllLinks() ([]*Link, error) {
 	}
 
 	return links, err
-}
-
-func calculateInitialFilterMask(i Interface) (mask byte) {
-
-	// Loopback devices will show all packets twice (for obvious reasons). In order to avoid
-	// this duplication, a default filter is applied, rejecting all outbound packets (this is
-	// in line with best practices, e.g. handling by libpcap).
-	if i.Type == TypeLoopback {
-		mask |= 4 // capture.PacketOutgoing
-	}
-
-	return
 }